requirements.txt
# Core Data Science
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<2.0.0
scipy>=1.7.0,<2.0.0

# Machine Learning
scikit-learn>=1.0.0,<2.0.0
umap-learn>=0.5.2,<0.6.0
joblib>=1.2.0,<2.0.0
nltk>=3.8.1,<4.0.0

# Web Application
streamlit>=1.10.0,<2.0.0
python-dotenv>=0.21.0,<1.0.0

# Data Visualization
matplotlib>=3.4.0,<4.0.0
plotly>=5.3.0,<6.0.0
wordcloud>=1.9.2,<2.0.0

# Text Processing
regex>=2021.11.0,<2023.0.0

# Utilities
tqdm>=4.62.0,<5.0.0

# Development Tools
jupyter>=1.0.0,<2.0.0
ipykernel>=6.0.0,<7.0.0

# Testing
pytest>=7.0.0,<8.0.0
pytest-cov>=3.0.0,<4.0.0

# Code Quality
black>=22.0.0,<23.0.0
flake8>=4.0.0,<5.0.0
isort>=5.10.0,<6.0.0
mypy>=0.910,<1.0.0

# Documentation
sphinx>=4.0.0,<5.0.0
sphinx-rtd-theme>=1.0.0,<2.0.0
myst-parser>=0.18.0,<0.19.0
